Add recording timer based on last frame timestamp
We could also easily implement this by having a timer run in the main
thread, but by consuming the last-recorded-frame timestamp, the time
displayed will also give immediate feedback about the current recording
status, which could be handy for detecting errors.
If this ever turns out to impact performance too much, we can still
easily change this to using a normal, second timer.