Don't resend settings when framerate is low
It never helped and caused a feedback loop where the system would never
recover (resending settings leads to a lower FPS measurement for the
current cycle, which again leads to resending settings and so on).
A previous revision fixed this behavior, but resending settings actually
doesn't appear to matter at all for low framerate issues, so we just
remove this condition (and a bit of additional complexity with it).