Simplify & improve time keeping code
Some of the changes are needed for better Syntalos integration, the
others just make the code easier to read and less complex. For example,
abusing time_point to keep times for a clock that it wasn't assigned to
was outright evil, and I am surpised this ever worked well enough.