Fix Python module build and switch to pybind11
This patch makes the Python module compile again, but it does not make
it usable yet, since some important methods to control the Miniscope
from Python are still missing.
This patch also switches from Boost::Python to pyBind11, which is much
nicer to use and supports PIMPL-using classes (it also compiles a bit
faster, which is nice for bigger projects).