Implement a device hard-reset feature for developers
This allows one to hard-reset the DAQ box completely from software. This
should not be needed at all in normal operation, but can be very useful
for developers or to get out of a situation where the DAQ box can not be
power-cycled manually (e.g. on remotely operated systems).