README: Add note on missing BNO info on newer Linux kernels
Unfortunately the PR to fix the firmware to advertise its controls
properly has still not been merged, and this error is really tricky for
users to debug.
The least we can do is to mention this in the README (and potentially in
future update the firmware version number to throw an actual warning,
sadly I didn't do that in the first place).